requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bert-extractive-summarizer==0.10.1
blis==1.0.1
breadability==0.1.20
catalogue==2.0.10
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
docopt==0.6.2
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.1
fsspec==2024.10.0
gensim==3.8.3
h11==0.14.0
httptools==0.6.4
huggingface-hub==0.26.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
packaging==24.1
preshed==3.0.9
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.40.0
summarization==0.0.1
sumy==0.11.0
sympy==1.13.1
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.2
urllib3==1.26.16
uvicorn==0.32.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==0.24.0
weasel==0.4.1
websockets==13.1
wrapt==1.16.0
